fix: modal styling on mobile (#179)
* fix: modal styling on mobile

* chore: update modal snapshots

When extracting the modals we missed the part that made them full-screen on mobile, they had full-width but not height anymore. Since all Enrique's design have modals full-height on mobile it led to quite an inconsistency with the intended design and how pages actually looked like.

This fixes the issue and restores full-screen modals on mobile.